+++ egui/src/containers/window.rs
@@ -720,7 +720,7 @@ impl TitleBar {
         self.title_label = self.title_label.text_color(style.fg_stroke.color);
 
         let full_top_rect = Rect::from_x_y_ranges(self.rect.x_range(), self.min_rect.y_range());
-        let text_pos = align::center_size_in_rect(self.title_galley.size, full_top_rect);
+        let text_pos = math::align::center_size_in_rect(self.title_galley.size, full_top_rect);
         let text_pos = text_pos.left_top() - 2.0 * Vec2::Y; // HACK: center on x-height of text (looks better)
         self.title_label
             .paint_galley(ui, text_pos, self.title_galley);