+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 * Center window titles.
 * Tweak size and alignment of some emojis to match other text.
+* Rename feature "serde" to "persistence".
 
 ### Fixed 🐛